refactor(label-filters): improve the label filtering
This removes --label and --ignore-label in favor of only using --labels
and --ignore-labels. Having a singular and plural version caused
confusion and created additional code checks that weren't needed.

It also create some new tests to check for validity of the new logic.

Note that it changes the prLabels from undefined []struct{} to []string.
Eventually it might go back to []PR once it's properly defined. But
until then we can use strings only.